fix: app switcher submenu
